More token optimizations
The limitation now doesn't lie in scanning the tokens, but rather in
getting the stack transitions.  This is where we had expected the
limitations to be.  It can be optimized fairly easily by creating an
N-dimensional array, rather than looping through lists but we're
probably better off by first looking at the other portions of what this
component needs to execute.