Rudimentary manipulation of a SPARQL query
This implementation is a hacked rudimentary SPARQL query generation
which would not be correct, but it would give us some insights into the
cost of query processing in Lisp.

The query processing we are taking into account is rudimentary changing
of the query, verifying it is still valid, and generating the query
results.  This allows us to make some estimations on the validity of the
work executed work.