More responsibilities for EBNF package
The EBNF package now transforms the supplied EBNF to an EBNF that cares
about casing of some of the core aspects.  The common structure of the
EBNF s-expressions has been converted into upper case symbols which
makes it more in line with Common Lisp's expectations.  This allows for
a bit cleaner processing further down the line.