Add types and type system, remove dead code
This is an extensive operation which removes dead code and adds advanced
types to the mix.  They can be disabled by lowering debug and safety
settings and they have a huge performance impact.  However, these
extensive checks can help us migrate to a faster approach in a more safe
manner.