First PoC for extracting quads to process
This only understands InsertData statements.  It includes a framework to
make the processing somewhat understandable.  The general idea being
that our general action is to pass on the parsing, to actively reject
some cases, as well as to allow some targeted processing.

This combines the best of the previous approach.  We want to consider
each option consciously and act on it as we should, but the vast
majority of actions is "let the level below handle it".  Passing
information is executed through a shared object in this case, which lets
us pass information around in a left-deep-first pass, and back up.

An example cobbled together test call is:

(sparql-parser:with-parser-setup
  (let ((detect-quads::*info* (detect-quads::make-info))
        (ast (sparql-parser:parse-sparql-string (coerce "INSERT DATA { GRAPH <http://foo.bar/> { <http://foo> <http://bar> <http://baz>. } }" 'base-string))))
    (sparql-parser:with-sparql-ast ast
      (format t "~&Processed sparql string:~%~A~%~A~%" sparql-parser::*scanning-string* (sparql-parser:sparql-ast-string ast))
      (princ

       (detect-quads-processing-handlers::|UpdateUnit|
                                          (sparql-parser:sparql-ast-top-node ast))))))