Many hacks in first steps towards processing update queries
Provides many hacks in a great and unclean commit dedicated in working
towards supporting update queries.

The following can now be executed, assuming nothing in the triplestore
will match the results :)

cl-user> (sparql-parser:with-parser-setup
           (let ((detect-quads::*info* (detect-quads::make-info))
                 (ast (sparql-parser:parse-sparql-string (coerce "PREFIX ex: <http://example.com/> INSERT DATA { GRAPH <http://foo.bar/> { <http://foo> <http://bar> <http://baz>. } }; DELETE DATA { GRAPH <http://example.com> { <http://example.com/> a <http://baz>, <http://bar>. } }; DELETE { GRAPH <http://foo.com> { ?s ex:myType ?type. } } INSERT { GRAPH <http://example.com> { ?s a ?type. } } WHERE { GRAPH <http://foo.bar> { ?s ex:myType ?type. } }; DELETE WHERE { GRAPH <http://foo.com> { ?s a <http://foo.com/Bar>, ?baz. } }" 'base-string))))
             (handle-update-unit::handle-sparql-update-unit (sparql-parser:sparql-ast-top-node ast))))

Further work should both include expanding on this functionality as well
as supporting delta messages.

The commit is rather ugly but keeping the state as a point in time is
important.