WIP towards updates
We have constructed a fine thread that implements the very primitves for
executing updates.  This includes continued steps towards that first
state for updating data.

This work introduces a test scenario which may help evaluate what works
and what does not.  This could later be converted into a less hands-on
test if that proves valuable.