Change name of EBNF NIL to SPARQLNIL
NIL is used within lisp to represent an empty list and the symbol is
self-evaluating.  This leads to sparql's nil to be the same as an empty
list which is not something expected during programming and may lead to
subtle bugs.  Renaming this symbol so it does not overlap should make
this more obvious.