Cope with selection when no graphs could be found
When no graphs could be found, the FROM portion would be empty, which in
turn would mean to select from all graphs.  That's quite the inverse
from what's intended.  We could construct a response directly but we
choose to select from an empty graph for now and let the backing sparql
endpoint figure out the ideal format for an empty response.