Supply scope of call to delta message
We currently assume the scope could both be used to apply more specific
access rights as well as to act more intelligently on delta messages
which should be processed.  This likely means some form of inheritance
will be needed in the future but we'll bootstrap it this way, knowing
the interface will change.