BASE and PREFIXES should not stick with UPDATE, extending TODO
This is assumed throughout the code we have written as it seems to be
the case in implementations we have observed.  The implementation is
incorrect though, and splitting up queries this way could be a
substantial endeavour in the stack.  We'll leave it as is but extend the
TODO with the necessary documentation.