Expand RDFLiteral in detect-quads so all IRIs are expanded
This should ensure all IRIs are expanded to our current knowledge.

During the implementation of this feature we've discovered that some
clients may need to alter their behaviour with respect to how BASE and
PREFIXES are being used, and perhaps this will require changes on
sparql-parser too.  The implementation provided here could be enough to
support current applications.