chore: log executed queries on failure
This logging is conditional on the same `*log-sparql-query-roundtrip*`
variable. The idea is that when someone is interested in in- and
outbound logs of a SPARQL query, they're probably also interested in
those logs if the query fails. If this ends up being too chatty, it can
be moved behind a separate variable.