Add a mechanism to replace large strings with URIs and files
Introduces as system which allows large strings to be replaced by URIs
and files backing these URIs.

The content is semi-transparent in that content is returned as strings.
However with the actual string not being in the SPARQL endpoint, some
querier (namely searching through the content of the string and sorting)
will operate based on the generated URI rather than on the string.

Even with that limitation in place, this seems to be a good step forward
for a bunch of applications where users store large strings into the
triplestore.