fix: reverse the PREFIXes when adding a BASE to the prologue
The AST of the SPARQL query stores the prefixes in reverse order for
some reason. This isn't necessarily the right place to fix this, but
it's an easy spot to reverse the reversal and ensure that the prefixes
are following the same order that the user supplied them in.

The order of the prefixes is important because if two prefixes are
defined, the latter "wins" and shadows the former. As of writing, the JS
template automatically adds some prefixes, one of which is mu with an
unexpected URI expansion. Without this fix, this ends up breaking
queries that depend on the mu prefix.