Only apply fix for prefixes in quad detection
The prefixes are stored in reverse order in the *info* object filled
when detecting quads.  The reverse order is faster and makes detection
of prefixes within the query faster.  However, the reverse order is
incorrect when generating the query for the WHERE clause.

An earlier fix reversed the order of the PREFIXES for all queries,
rather than only those of the update.  This code applies the change only
where necessary.

co-author: @erikap