Support scanning of empty PropertyListPathNotEmpty
The generated BNF did not yield the _empty sequence for
PropertyListPathNotEmpty.  This is necessary however, because this
contains a * pattern.  We have created an intermediate rule which
accepts _eps and yields an alternative of _empty and the original rule.

This problem was signaled by @sergiofenoll because the parser did not
parse `FILTER NOT EXISTS { [] a ex:Broken }`.  @sergiofenoll also
validated this to be valid sparql.  The problem was confirmed by
@ValenberghsSven to be of practical importance.

Co-author: @erikap