Wake delta message bus immediately on messages
Introduces a semaphore which will wake the message bus when a new
message is put on the queue.  An optional timeout can be set (currently
set to 60 seconds) to indicate how long we want to wait when there are
no messages.  This is mostly to circumvent undiscovered race conditions
and it could be set to nil too.