Support sudo as mu-auth-allowed-group
Some services will send the mu-auth-allowed-group named "sudo" to
indicate it's a sudo call.  This also lets us pass the allowed group
through the delta notifier to consuming services so it somewhat makes
sense.

This behaviour should be considered subject to change though.  It can be
cleaned up to offer more structured functionality later on.