Add check for missing prefixes during prefix expansion
Users may forget to add prefixes.  In these cases we assume the user
supplied a protocol.  But a protocol such as foaf: or rdf: would be very
weird indeed.  There is now a check which checks that a missing prefix
has some somewhat sensible protocol instead.  It's user-extensible if
needed and the check can be disabled.

Co-author: @erikap