Make quad list a struct
This approach will allow us to store more information with the quad and
it will allow us to better know what is in the quad.  Further, this may
open up paths to smaller quad objects if we learn to understand what
their contents tend to be.

Code should be validated and tested.