Reuse terminal-match-string from sparql-parser
We had redefined this as primitive-match-string which may sound a bit
better but it's probably best to consume content from sparql-parser
instead.  This means we don't have a redifinition today.  We might
rename it later and that might break user configuration at that time but
it should be sane as it stands.