Support users in providing quad transformations
The following would verify that all quads which try to talk about a
mu:uuid with a datatype, are written as raw strings:

    (in-package :quad-transformations)

    (quad-transformations:define-quad-transformation (quad method)
      ;; make quad objects which have datatype in uuid specification just strings
      (if (and
           ;; predicate is uuid
           (string= (detect-quads::quad-term-uri (quad:predicate quad))
                    "http://mu.semte.ch/vocabularies/core/uuid")
           ;; object has datatype
           (= (length (sparql-parser:match-submatches (quad:object quad))) 3))
          (let ((new-quad (quad:copy quad))) ; make new quad
            (setf (quad:object new-quad)
                  (handle-update-unit::make-nested-match
                   `(ebnf::|RDFLiteral| ,(first (sparql-parser:match-submatches (quad:object quad))))))
            ;; use the new quad
            (quad-transformations:update new-quad))
          ;; otherwise keep it
          (quad-transformations:keep)))

The macro is not as hygienic as we'd want it to be but it acted up and
this would be a usable solution already.

This functionality is tied to inner workings of sparql-parser which are
subject to change.  This would allow for some experimentation and
band-aids.