Make sure /data folder is mounted at boot time
When sparql-parser runs without mounting /data, long strings are not
persisted to disk.  This may cause data loss.  Strings longer than 4k
characters are written to disk instead of to the triplestore, removing
the container will thus mean losing that data.

We now exit when the data folder is not mounted so this doesn't happen
in practice.

Co-author: @erikap