Support long strings with non-ascii characters
When a string contains characters which don't have a byte representation
in UTF-8, the string is not a byte vector.  The sha1 library accepts
strings but only if they are byte vectors.  We use flexi-streams to
convert the stream to an octet-string which should allow us to have
extended characters in long strings similar to how the triplestore would
support it.