Add singleton flight check system
This flight check system can be used to only allow certain tokens to be
shared.  This should allow us to ensure we are not sending overlapping
updates to the backend and should allow us to provide a rough order of
events to services which depend on the order of changes having been
written to the backend.