Increase max query time once first part of the query has been completed
Set `*max-query-time-for-retries-in-followup-queries*` for the amount of time
the next query should get and `*max-query-time-for-retries*` for the first try.

This still has intrinsic downsides.  An ideal solution would keep more state
about the sort of queries which work out and the sort of queries which do not
work out.  It may be that we want to cancel out some of the updates we've sent
beforehand.  There's a limit to how much we can understand of the endpoint not
responding because it's overloaded (and this coming through later) or the
request itself being too heavy to lift.