Split up acl:dispatch-quads
The function was responsible for detecting the relevant graph specifications as
well as distributing the quads.  As we want to progress whether quads have to be
written and where they should be written, we're splitting off the moving of
quads.

Splitting off the moving of quads should allow us to better determine the
execution of what we should do with the quads.  Our hope is that this will also
allow us to decide more intelligenty what to do with the quads.  Whether we want
to write the quad operations to the triplestore or to a file, or whether we want
to write the quads at all (and considering them written for the
error-on-unwritten-data case).

This work in turn fits within the larger operation of push updates.  These
improvements should support the generation of delta messages _without_ requiring
the quads to be written.  We want such things to allow microservices to write
more data to the triplestore as information (eg: warnings or information
elements).