Extend graph configuration interface and split up code
This doesn't look like much but it has split up the code which understands which
quads are treated by which graph+token from the code that moves the quads.  We
can now connect a quad to the corresponding graph+token.  Further, an interface
has been defined which emits many many values which should be used for further
handling the delta messages and writing data.

The graph configuration interface now takes a few options, but these are not
used yet.  The idea is that these options will allow us to indicate that certain
data needs to be written to the triplestore and/or should emit delta messages.
This is not complete by far and it's also not the end of the desired features.
But it's a solid start nonetheless.