Support graphs which will only cause delta messages but not sparql
The graph definitions can now receive :sparql nil and :delta nil to disable
certain aspects of writing to those graphs.  Not having SPARQl means there are
no effective changes to be emitted but you can still have the messages emit
other data.  This may help for systems which only want to act on the delta
messages without having the data written to the sparql endpoint.