Ensure mu-auth-sudo is applied
The rules for mu-auth-sudo were not applied correctly, leading to warnings about
unwritten data.

This implementation writes all data to the triplestore which is the standard way
of working just about everywhere.  This may change in the future as we may want
to abide how to treat graphs even when making a sudo query so we can still use
the controls the graphs have specified maybe.  By the time we implement such
features we will likely have to provide workarounds too.