Clean up sparql results received from Virtuoso
Virtuoso sometimes supplies numbers as numbers instead of strings in the sparql
result body.  That's not desired because we may cast these into a ration or a
double and when printed readably by lisp, they're unreadable in the SPARQL
syntax.  This commit tries to clean up the writing of numbers so they get
printed in a manner which would work for a sparql result set and can reliably be
converted back into SPARQL.