Allow OPTIONAL in CONSTRUCT for MODIFY to we written as UNION
A MODIFY query can be rewritten to use a CONSTRUCT rather than a SELECT to
discover the data to be written in some cases.  This CONSTRUCT may contain
OPTIONAL statements which is valid, yet sometimes problematic.  Specifically in
Virtuoso this may yield issues.  This variable being T means we'll rewrite the
OPTIONAL so it becomes a UNION with an empty accepted clause.

This solution should be verified when various parts of the query are missing.
This might be different between versions and perhaps it changes based on the
available data.